// src/Models/Location_Search.ts
export default class Location_Search{
  "place_id": number;
  "licence": string;
  "osm_type": string;
  "osm_id": number;
  "lat": string;
  "lon": string;
  "category": string;
  "type": string;
  "place_rank": number;
  "importance": number;
  "addresstype": string;
  "name": string;
  "display_name": string;
  "boundingbox": string[];
  "geojson": GeoJson;
}
export class GeoJson{
  "type": string;
  "coordinates": number[][][];
}